renaming calls to provider in articles js.
